+++ openeo/rest/auth/auth.py
@@ -46,12 +46,8 @@ class BasicBearerAuth(BearerAuth):
         super().__init__(bearer='basic//{t}'.format(t=access_token))
 
 
-OidcRefreshInfo = collections.namedtuple("_OidcRefreshData", ["provider_id", "client_id"])
-
-
 class OidcBearerAuth(BearerAuth):
     """Bearer token for OIDC Auth (openEO API 1.0.0 style)"""
 
-    def __init__(self, provider_id: str, access_token: str, refresh_data: Optional[OidcRefreshInfo] = None):
+    def __init__(self, provider_id: str, access_token: str):
         super().__init__(bearer='oidc/{p}/{t}'.format(p=provider_id, t=access_token))
-        self.refresh_data = refresh_data